refactor: remove and refactor unused old selfkey-id components
