--- src/renderer/selfkey-id/main/components/selfkey-id.jsx
@@ -1,27 +1,0 @@
-import React, { PureComponent } from 'react';
-import { Grid, Typography, Tabs, Tab } from '@material-ui/core';
-
-class SelfkeyIdComponent extends PureComponent {
-	render() {
-		return (
-			<Grid container direction="column" spacing={4}>
-				<Grid item>
-					<Typography variant="h1">SelfKey Profile</Typography>
-				</Grid>
-				<Grid item>
-					<Tabs value={this.props.tab} onChange={this.props.onTabChange}>
-						<Tab label="Overview" />
-						<Tab label="Marketplace Applications" />
-						{/* <Tab label="Companies" /> */}
-						{/* <Tab label="History" /> */}
-					</Tabs>
-				</Grid>
-				<Grid item>{this.props.component}</Grid>
-			</Grid>
-		);
-	}
-}
-
-export const SelfkeyId = SelfkeyIdComponent;
-
-export default SelfkeyId;
