--- src/renderer/selfkey-id/main/index.js
@@ -1,1 +1,0 @@
-export { SelfkeyIdContainer } from './containers/selfkey-id';
